Update optimizations.md
Using `strip = true` works fine for everything on Mac but on Linux the stripped binary can't be used to generate. `strip = 'debuginfo'` gets most of the way there but can still be used for generating, and can be stripped the rest of the way with the Unix tool `strip` if desired.

If this is _only_ intended for use when creating the binary sent out with the packaged result, and can't be used for generating using `--library`, it might be worth calling out.